# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hệ Thống Tự Động Sinh Test Case API (CDIO-4)",
    "description": "Nền tảng tự động bóc tách đặc tả yêu cầu, giải ràng buộc và sinh bộ kiểm thử phần mềm đa phương pháp (BVA, Phân vùng tương đương, Pairwise, Z3 Solver). Quản lý chu kỳ Scrum, phân công nhiệm vụ thành viên và xuất báo cáo kiểm thử nghiệm thu chuyên nghiệp.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/sprints/project/{project_id}": {
      "get": {
        "tags": [
          "Quản lý Sprint & Scrum"
        ],
        "summary": "Danh sách Sprint của dự án",
        "description": "Lấy danh sách toàn bộ các chu kỳ Sprint trong dự án kèm chỉ số hoàn thành tiêu chuẩn Definition of Done (DoD).",
        "operationId": "get_project_sprints_api_sprints_project__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SprintOut"
                  },
                  "title": "Response Get Project Sprints Api Sprints Project  Project Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sprints": {
      "post": {
        "tags": [
          "Quản lý Sprint & Scrum"
        ],
        "summary": "Tạo Sprint mới trong dự án",
        "description": "Khởi tạo một chu kỳ Sprint mới với mục tiêu kiểm thử (Sprint Goal), ngày bắt đầu và kết thúc.",
        "operationId": "create_sprint_api_sprints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SprintCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sprints/{sprint_id}": {
      "put": {
        "tags": [
          "Quản lý Sprint & Scrum"
        ],
        "summary": "Cập nhật thông tin Sprint",
        "description": "Chỉnh sửa tên Sprint, mục tiêu, ngày bắt đầu, ngày kết thúc hoặc trạng thái hoàn thành.",
        "operationId": "update_sprint_api_sprints__sprint_id__put",
        "parameters": [
          {
            "name": "sprint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sprint Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SprintUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Quản lý Sprint & Scrum"
        ],
        "summary": "Xóa Sprint",
        "description": "Xóa chu kỳ Sprint và gỡ bỏ liên kết của các chức năng con thuộc Sprint này.",
        "operationId": "delete_sprint_api_sprints__sprint_id__delete",
        "parameters": [
          {
            "name": "sprint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sprint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SprintCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "goal": {
            "type": "string",
            "title": "Goal",
            "default": ""
          },
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "default": ""
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "project_id"
        ],
        "title": "SprintCreate"
      },
      "SprintOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "goal": {
            "type": "string",
            "title": "Goal",
            "default": ""
          },
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "default": ""
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "requirements_count": {
            "type": "integer",
            "title": "Requirements Count",
            "default": 0
          },
          "total_cases": {
            "type": "integer",
            "title": "Total Cases",
            "default": 0
          },
          "pass_count": {
            "type": "integer",
            "title": "Pass Count",
            "default": 0
          },
          "fail_count": {
            "type": "integer",
            "title": "Fail Count",
            "default": 0
          },
          "pass_rate": {
            "type": "number",
            "title": "Pass Rate",
            "default": 0.0
          },
          "dod_met": {
            "type": "boolean",
            "title": "Dod Met",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "project_id"
        ],
        "title": "SprintOut"
      },
      "SprintUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "SprintUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}